Add example 5.1 seek target. Added DrawLines method.

+++ Chapter03/Example_03_02_Mover.cpp
@@ -30,7 +30,7 @@ namespace nature_of_code_chapter_03_example_02
         int line_x2 = radius_;
         int line_y2 = 0;
 
-        framework->Translate((int)position_.x, (int)position_.y);
+        framework->Translate(position_.x, position_.y);
         framework->RotateRadians(angle_);
         framework->DrawCircle(0, 0, radius_ * 2, kColorGray, true);
         framework->DrawLine(line_x1, line_y1, line_x2, line_y2, kColorBlack);
